- Maintain cache of ArrowFlightClient
